+++ test/unit/fuzz-tools.test.js
@@ -299,6 +299,7 @@ import * as ccsAnginaV319 from '../../lib/ccs-angina-v319.js';
 import * as clavienDindoV320 from '../../lib/clavien-dindo-v320.js';
 import * as hincheyV321 from '../../lib/hinchey-v321.js';
 import * as biRadsV322 from '../../lib/bi-rads-v322.js';
+import * as siewertV323 from '../../lib/siewert-v323.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -543,6 +544,7 @@ const MODULES = {
   'clavien-dindo-v320.js': clavienDindoV320,
   'hinchey-v321.js': hincheyV321,
   'bi-rads-v322.js': biRadsV322,
+  'siewert-v323.js': siewertV323,
   'tb-testing.js': tbTesting,
 };